In progress refactoring of post processors into its own package and sorting out demographic groups
