+++ safe/definitions/field_groups/gender_vulnerability_field_group.py
@@ -29,7 +29,7 @@ __email__ = "info@inasafe.org"
 __revision__ = '$Format:%H$'
 
 gender_vulnerability_ratio_group = {
-    'key': 'gender_ratio_group',
+    'key': 'gender_vulnerability_ratio_group',
     'name': tr('Gender Vulnerability Ratio'),
     'description': tr(
         'Demographic breakdown to use for displaced population based on '
@@ -64,7 +64,7 @@ gender_vulnerability_ratio_group = {
 }
 
 gender_vulnerability_count_group = {
-    'key': 'gender_count_group',
+    'key': 'gender_vulnerability_count_group',
     'name': tr('Gender Vulnerability Count'),
     'description': tr(
         'Demographic breakdown to use for displaced population based on '
@@ -90,7 +90,7 @@ gender_vulnerability_count_group = {
 }
 
 gender_vulnerability_displaced_count_group = {
-    'key': 'gender_displaced_count_group',
+    'key': 'gender_vulnerability_displaced_count_group',
     'name': tr('Gender Vulnerability Displaced Count'),
     'description': tr(
         'Demographic breakdown to use for displaced population based on '
