+++ safe/impact_function/test/test_postprocessors.py
@@ -23,18 +23,16 @@ from safe.definitions.fields import (
     hygiene_packs_count_field,
     additional_rice_count_field)
 from safe.definitions.post_processors import (
-    post_processor_gender,
-    post_processor_youth,
-    post_processor_adult,
-    post_processor_elderly,
     post_processor_size_rate,
     post_processor_size,
     post_processor_affected,
     field_input_type,
     dynamic_field_input_type,
     needs_profile_input_type,
-    post_processor_hygiene_packs,
     post_processor_additional_rice)
+from safe.definitions.post_processors.population_post_processors import \
+    post_processor_gender, post_processor_hygiene_packs, post_processor_youth, \
+    post_processor_adult, post_processor_elderly
 from safe.test.utilities import load_test_vector_layer
 from safe.impact_function.postprocessors import (
     run_single_post_processor,
